build: improve building workflow
This change improves the building workflow by putting the platforms
management in a central place. The platforms definition schema is the
same used by docker and then is splitted to be used the same schema in
the Go builds as well.

This change also brings an experimental multi-arch building by using
`buildx`.